requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity
azure-storage-blob==2.1.0
psycopg2-binary
pytest
requests
multiweatherapi==0.0.16
opencensus-ext-azure